[Everyone on dev should pull this!] Library Reorganization.
