+++ examples/move_with_no_borrow/annotated_source.rs
