+++ examples/move_with_no_borrow/source.rs
