+++ examples/report_eg/example.rs
