+++ examples/simple_lifetime/example2.rs
